# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GoaGuide API
  description: Production-grade travel itinerary & marketplace platform for Goa
  version: 1.0.0
  contact:
    name: GoaGuide API Support
    email: api-support@goaguide.com
  license:
    name: Proprietary
    url: https://goaguide.com/license
servers:
  - url: https://api.goaguide.com/v1
    description: Production server
  - url: https://staging-api.goaguide.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /trips:
    post:
      tags:
        - Trips
      summary: Create a new trip
      description: Create a new trip with minimal input for itinerary generation
      operationId: createTrip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTripRequest'
            examples:
              minimal:
                summary: Minimal trip creation
                value:
                  destination: "Goa"
                  message: "I want to go to Goa"
              detailed:
                summary: Detailed trip creation
                value:
                  destination: "Goa"
                  message: "Family trip to Goa with kids"
                  dates:
                    start_date: "2024-12-15"
                    end_date: "2024-12-17"
                  party_composition:
                    adults: 2
                    children: 2
                    infants: 0
                  budget_per_person: 5000
                  preferences:
                    - "beaches"
                    - "family-friendly"
      responses:
        '201':
          description: Trip created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /events:
    get:
      tags:
        - Events
      summary: Search local events
      description: Find events by location, date, and category
      operationId: searchEvents
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
            format: float
          example: 15.2993
        - name: lng
          in: query
          required: true
          schema:
            type: number
            format: float
          example: 74.1240
        - name: radius
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
          description: Search radius in kilometers
        - name: date
          in: query
          schema:
            type: string
            format: date
          example: "2024-12-15"
        - name: category
          in: query
          schema:
            type: string
            enum: [festival, market, party, cultural, sports, food]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Events retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventsResponse'

  # Provider Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    ProviderApiKey:
      type: apiKey
      in: header
      name: X-Provider-API-Key
...
  schemas:
    CreateTripRequest:
      type: object
      required:
        - destination
        - message
      properties:
        destination:
          type: string
          example: "Goa"
        message:
          type: string
          example: "I want to go to Goa for a family vacation"
        dates:
          type: object
          properties:
            start_date:
              type: string
              format: date
            end_date:
              type: string
              format: date
        party_composition:
          type: object
          properties:
            adults:
              type: integer
              minimum: 1
            children:
              type: integer
              minimum: 0
            infants:
              type: integer
              minimum: 0
        budget_per_person:
          type: number
          minimum: 500
        preferences:
          type: array
          items:
            type: string
    TripResponse:
      type: object
      properties:
        trip_id:
          type: string
        status:
          type: string
          enum: [created, questions_pending, generating, ready, booked]
        destination:
          type: string
        follow_up_questions:
          type: array
          items:
            $ref: '#/components/schemas/FollowUpQuestion'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FollowUpQuestion:
      type: object
      properties:
        question_id:
          type: string
        question:
          type: string
        type:
          type: string
          enum: [single_choice, multiple_choice, text, number, date]
        options:
          type: array
          items:
            type: string
        required:
          type: boolean
...
    Location:
      type: object
      properties:
        lat:
          type: number
          format: float
        lng:
          type: number
          format: float
        address:
          type: string
        poi_id:
          type: string
...
    EventsResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        total_count:
          type: integer
        has_more:
          type: boolean
    Event:
      type: object
      properties:
        event_id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        location:
          $ref: '#/components/schemas/Location'
        price_range:
          type: object
          properties:
            min:
              type: number
            max:
              type: number
        confidence_score:
          type: number
          minimum: 0
          maximum: 1
        source:
          type: string
        impact_level:
          type: string
          enum: [low, medium, high]
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            trace_id:
              type: string
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "INVALID_INPUT"
              message: "Required field 'destination' is missing"
              trace_id: "trace_123456"
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Invalid or expired JWT token"
              trace_id: "trace_123456"
...
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "RATE_LIMITED"
              message: "Rate limit exceeded. Try again in 60 seconds"
              trace_id: "trace_123456"
      headers:
        Retry-After:
          schema:
            type: integer
          description: Number of seconds to wait before retrying